Convert from the given Preorder Traversal to a BST
Pre order traversal is given and we need to convert it to a BST
Approach one is to find the inorder traversal by just sorting the preorder traversal given

Then do the inorderToBST conversion
But this takes O(nxn) time.

So a better approach will be-

Approach #2
We can convert directly from Pre order to BST
Using the property of range. INT_MAX and INT_MIN use